requirements.txt
PyYAML
requests
playsound3
pytz
schedule